file add & rm ready

+++ main.c
@@ -129,21 +129,20 @@ int mKKdir(KK failsisten32, char* name){
 	
 int rmKKdir(KK failsisten32, char* name){
 	/*	Remueve un directorio en el nodo actual de la tabla, 
-	si no se encuentra el archivo recibimos -1 */
-	
+	si no se encuentra el archivo recibimos -1 */	
 
 	if( failsisten32.curr->child == NULL ){
 		// no hay dirs
 		printf("No existe el directorio. \n");
 		return -1;
 
-	} else if( failsisten32.curr->child->name == name){
+	} else if( failsisten32.curr->child->name == name && failsisten32.curr->child->type == 1 ){
 		// es el primer dir
 		int res = rmdir( failsisten32.curr->child->path );
 		if(res < 0){
 		  printf("No se pudo borrar el directorio\n");
 		  return -1;
-
+		}
 		// Manejar archivos
 		NODO* temp = failsisten32.curr->child->next;
 		printf("Removiendo: %s\n", failsisten32.curr->child->path);
@@ -162,7 +161,7 @@ int rmKKdir(KK failsisten32, char* name){
 
 		NODO* temp = failsisten32.curr->child;
 		while(temp->next != NULL){
-			if( temp->next->name == name){
+			if( temp->next->name == name && temp->child->type == 1){
 				NODO* new = temp->next->next;
 				free( temp->next );	
 				temp->next = new;
@@ -171,6 +170,7 @@ int rmKKdir(KK failsisten32, char* name){
 		}
 		return -1;
 	}
+
 }
 
 int touchWC(KK failsisten32, char* name){
@@ -215,4 +215,50 @@ int touchWC(KK failsisten32, char* name){
 	}	
 
 	return 0;
+}
+
+int rmWC(KK failsisten32, char* name){
+	/*	Remueve un directorio en el nodo actual de la tabla, 
+	si no se encuentra el archivo recibimos -1 */	
+
+	if( failsisten32.curr->child == NULL ){
+		// no hay dirs
+		printf("No existe el directorio. \n");
+		return -1;
+
+	} else if( failsisten32.curr->child->name == name && failsisten32.curr->child->type == 0 ){
+		// es el primer dir
+		int res = rmdir( failsisten32.curr->child->path );
+		if(res < 0){
+		  printf("No se pudo borrar el directorio\n");
+		  return -1;
+		}
+		// Manejar archivos
+		NODO* temp = failsisten32.curr->child->next;
+		printf("Removiendo: %s\n", failsisten32.curr->child->path);
+		free( failsisten32.curr->child );	
+		failsisten32.curr->child = temp;
+	}else{
+		// es otro dir
+		int res = rmdir( failsisten32.curr->child->path );
+
+		printf("Removiendo: %s\n", failsisten32.curr->child->path );
+		if(res < 0){
+			printf("No se ha podido borrar el directorio.\n");
+			return -1;
+		}
+		// Manejar archivos
+
+		NODO* temp = failsisten32.curr->child;
+		while(temp->next != NULL){
+			if( temp->next->name == name && temp->child->type == 0){
+				NODO* new = temp->next->next;
+				free( temp->next );	
+				temp->next = new;
+				return 0;
+			}
+		}
+		return -1;
+	}
+
 }
